Evaluate trig in double, because the client's CRT does
Replaying a PC recording on Android diverges at the first checksum checkpoint,
frame 100, on an idle map -- and of 317 objects exactly 11 disagree: ten
civilian ambient vehicles and a dozer. Every static object hashes byte
identical. Only the things that MOVE diverge.

The movement code itself is not the difference. Locomotor, PhysicsUpdate,
AIUpdate, AIStates, AIPathfind, TerrainLogic, ScriptEngine and GameLogic were
all diffed against the PC client's own source and are clean, and Locomotor and
PhysicsUpdate turn out to declare no doubles at all. Nor is it the
architecture: aarch64 and x86_64 produce bit-identical results for the whole
transcendental set in single and double precision.

It is which libm each side links. The source here is character-identical to
the client's -- plain sinf/cosf -- but the client is VC6 for 32-bit x86, whose
CRT promotes to double and evaluates on x87, while bionic implements a real
single-precision sinf/cosf with its own error. Measured against the client's
exact contract, x87 with the control word setFPMode() installs (PC=24,
round-to-nearest), over 62801 angles spanning [-pi, pi]:

    x87 fsin  vs  sinf                      776 differ  (1.236%)
    x87 fcos  vs  cosf                      768 differ  (1.223%)
    x87 fsin  vs  (float)sin((double)x)       0 differ
    x87 fcos  vs  (float)cos((double)x)       0 differ

One angle in eighty. A rotation matrix is rebuilt for every moving object
every frame and Object::crc hashes the transform directly, so a static scene
stays in step and a driving car does not -- which is the observed signature,
exactly.

So evaluate in double and narrow once, at all 42 sites in wwmath.h, matrix3d.h
and matrix3.h. This is the pattern the same files already use for Acos, Asin
and Sqrt, and the MSVC/x86 x87 branches are untouched, so a VC6 build still
takes the client's own path. It states in the source the precision contract
the client gets implicitly from its CRT.

+++ Core/Libraries/Source/WWVegas/WWMath/matrix3d.h
@@ -564,10 +564,40 @@ WWINLINE void Matrix3D::Set(		const Vector3	&x,		// x-axis unit vector
  * HISTORY:                                                                                    *
  *   2/24/98    GTH : Created.                                                                 *
  *=============================================================================================*/
+/*
+ * GeneralsX @bugfix Android port 20/09/2026 -- cross-play determinism.
+ *
+ * These call sinf/cosf/tanf, and the source is character-identical to the PC
+ * GeneralsOnline client's. The divergence is not in the code, it is in which
+ * libm the two sides link. The client is built with VC6 for 32-bit x86, whose
+ * CRT promotes the argument to double and evaluates on the x87 unit; bionic
+ * implements a genuine single-precision sinf/cosf with its own ~1 ULP error.
+ * They land on different floats.
+ *
+ * Measured against the client's exact contract -- x87 with the control word
+ * setFPMode() installs (PC=24, round-to-nearest) -- over 62801 angles across
+ * [-pi, pi]:
+ *
+ *     x87 fsin  vs  sinf                      776 differ  (1.236%)
+ *     x87 fcos  vs  cosf                      768 differ  (1.223%)
+ *     x87 fsin  vs  (float)sin((double)x)       0 differ  (0.000%)
+ *     x87 fcos  vs  (float)cos((double)x)       0 differ  (0.000%)
+ *
+ * One angle in eighty is enough: a rotation matrix is rebuilt for every moving
+ * object every frame, and Object::crc hashes the transform directly. It is
+ * also exactly the observed signature -- replaying a PC recording on Android
+ * diverges at the first checkpoint with only the objects that MOVE disagreeing
+ * (11 of 317: ten civilian vehicles and a dozer), every static object hashing
+ * byte-identical.
+ *
+ * So evaluate in double and narrow once, which is what this same file already
+ * does for Acos, Asin and Sqrt. Not a workaround: it states in the source the
+ * precision contract that the client gets implicitly from its CRT.
+ */
 WWINLINE void Matrix3D::Set(const Vector3 & axis,float angle)
 {
-	float c = cosf(angle);
-	float s = sinf(angle);
+	float c = (float)cos((double)(angle));
+	float s = (float)sin((double)(angle));
 
 	Set(axis,s,c);
 }
@@ -768,8 +798,8 @@ WWINLINE void Matrix3D::Rotate_X(float theta)
 	float tmp1,tmp2;
 	float s,c;
 
-	s = sinf(theta);
-	c = cosf(theta);
+	s = (float)sin((double)(theta));
+	c = (float)cos((double)(theta));
 
 	tmp1 = Row[0][1]; tmp2 = Row[0][2];
 	Row[0][1] = (float)( c*tmp1 + s*tmp2);
@@ -836,8 +866,8 @@ WWINLINE void Matrix3D::Rotate_Y(float theta)
 	float tmp1,tmp2;
 	float s,c;
 
-	s = sinf(theta);
-	c = cosf(theta);
+	s = (float)sin((double)(theta));
+	c = (float)cos((double)(theta));
 
 	tmp1 = Row[0][0]; tmp2 = Row[0][2];
 	Row[0][0] = (float)(c*tmp1 - s*tmp2);
@@ -903,8 +933,8 @@ WWINLINE void Matrix3D::Rotate_Z(float theta)
 	float tmp1,tmp2;
 	float c,s;
 
-	c = cosf(theta);
-	s = sinf(theta);
+	c = (float)cos((double)(theta));
+	s = (float)sin((double)(theta));
 
 	tmp1 = Row[0][0]; tmp2 = Row[0][1];
 	Row[0][0] = (float)( c*tmp1 + s*tmp2);
@@ -1055,8 +1085,8 @@ WWINLINE void Matrix3D::Pre_Rotate_X(float theta)
 	float tmp1,tmp2;
 	float c,s;
 
-	c = cosf(theta);
-	s = sinf(theta);
+	c = (float)cos((double)(theta));
+	s = (float)sin((double)(theta));
 
 	tmp1 = Row[1][0]; tmp2 = Row[2][0];
 	Row[1][0] = (float)(c*tmp1 - s*tmp2);
@@ -1093,8 +1123,8 @@ WWINLINE void Matrix3D::Pre_Rotate_Y(float theta)
 	float tmp1,tmp2;
 	float c,s;
 
-	c = cosf(theta);
-	s = sinf(theta);
+	c = (float)cos((double)(theta));
+	s = (float)sin((double)(theta));
 
 	tmp1 = Row[0][0]; tmp2 = Row[2][0];
 	Row[0][0] = (float)( c*tmp1 + s*tmp2);
@@ -1131,8 +1161,8 @@ WWINLINE void Matrix3D::Pre_Rotate_Z(float theta)
 	float tmp1,tmp2;
 	float c,s;
 
-	c = cosf(theta);
-	s = sinf(theta);
+	c = (float)cos((double)(theta));
+	s = (float)sin((double)(theta));
 
 	tmp1 = Row[0][0]; tmp2 = Row[1][0];
 	Row[0][0] = (float)(c*tmp1 - s*tmp2);
@@ -1274,8 +1304,8 @@ WWINLINE void Matrix3D::In_Place_Pre_Rotate_X(float theta)
 	float tmp1,tmp2;
 	float c,s;
 
-	c = cosf(theta);
-	s = sinf(theta);
+	c = (float)cos((double)(theta));
+	s = (float)sin((double)(theta));
 
 	tmp1 = Row[1][0]; tmp2 = Row[2][0];
 	Row[1][0] = (float)(c*tmp1 - s*tmp2);
@@ -1308,8 +1338,8 @@ WWINLINE void Matrix3D::In_Place_Pre_Rotate_Y(float theta)
 	float tmp1,tmp2;
 	float c,s;
 
-	c = cosf(theta);
-	s = sinf(theta);
+	c = (float)cos((double)(theta));
+	s = (float)sin((double)(theta));
 
 	tmp1 = Row[0][0]; tmp2 = Row[2][0];
 	Row[0][0] = (float)( c*tmp1 + s*tmp2);
@@ -1342,8 +1372,8 @@ WWINLINE void Matrix3D::In_Place_Pre_Rotate_Z(float theta)
 	float tmp1,tmp2;
 	float c,s;
 
-	c = cosf(theta);
-	s = sinf(theta);
+	c = (float)cos((double)(theta));
+	s = (float)sin((double)(theta));
 
 	tmp1 = Row[0][0]; tmp2 = Row[1][0];
 	Row[0][0] = (float)(c*tmp1 - s*tmp2);
